KDF.PBKDF2 protection from brute force

+++ fusecry/filesystem.py
@@ -4,10 +4,9 @@ Fusecry FUSE operations.
 
 from datetime import datetime
 from fuse import FuseOSError, Operations
-from fusecry import cry
-from fusecry.io import FusecryIO
+from fusecry import config
+from fusecry import io
 import errno
-import fusecry.config as config
 import os
 
 def debug_log(func):
@@ -27,33 +26,41 @@ class Fusecry(Operations):
     def __init__(self, root, password, ignore_ic=False, debug=False):
         self.root = root
         self.debug = debug
-        self.io = FusecryIO(cry.Cry(password), ignore_ic)
+        self.conf = os.path.join(self.root, config.enc.conf)
+        self.io = io.make_io(password, self.conf, ignore_ic)
 
-    def __real_path(self, partial):
-        if partial.startswith("/"):
-            partial = partial[1:]
-        path = os.path.join(self.root, partial)
-        return path
+    def __real_path(self, path):
+        if path.startswith(os.path.sep):
+            path = path[len(os.path.sep):]
+        return os.path.join(self.root, path)
 
     # Filesystem methods
     # ==================
     
     @debug_log
     def access(self, path, mode):
-        if not os.access(self.__real_path(path), mode):
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None
+        if not os.access(real_path, mode):
             raise FuseOSError(errno.EACCES)
     
     @debug_log
     def chmod(self, path, mode):
-        return os.chmod(self.__real_path(path), mode)
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None
+        return os.chmod(real_path, mode)
     
     @debug_log
     def chown(self, path, uid, gid):
-        return os.chown(self.__real_path(path), uid, gid)
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None
+        return os.chown(real_path, uid, gid)
     
     @debug_log
     def getattr(self, path, fh=None):
-        return self.io.attr(self.__real_path(path))
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None
+        return self.io.attr(real_path)
     
     @debug_log
     def readdir(self, path, fh):
@@ -61,12 +68,16 @@ class Fusecry(Operations):
         dirents = ['.', '..']
         if os.path.isdir(real_path):
             dirents.extend(os.listdir(real_path))
+            if os.path.abspath(real_path) == os.path.abspath(self.root):
+                if config.enc.conf in dirents:
+                    dirents.remove(config.enc.conf)
         for r in dirents:
             yield r
     
     @debug_log
     def readlink(self, path):
         real_path = self.__real_path(path)
+        if real_path == self.conf: return None
         pathname = os.readlink(real_path)
         if pathname.startswith("/"):
             # Path name is absolute, sanitize it.
@@ -77,21 +88,25 @@ class Fusecry(Operations):
     @debug_log
     def mknod(self, path, mode, dev):
         real_path = self.__real_path(path)
+        if real_path == self.conf: return None
         return os.mknod(self.real_path, mode, dev)
     
     @debug_log
     def rmdir(self, path):
         real_path = self.__real_path(path)
+        if real_path == self.conf: return None
         return os.rmdir(real_path)
     
     @debug_log
     def mkdir(self, path, mode):
         real_path = self.__real_path(path)
+        if real_path == self.conf: return None
         return os.mkdir(real_path, mode)
     
     @debug_log
     def statfs(self, path):
         real_path = self.__real_path(path)
+        if real_path == self.conf: return None
         stv = os.statvfs(real_path)
         stat = dict((key, getattr(stv, key)) for key in (
             'f_bavail', 'f_bfree', 'f_blocks', 'f_bsize', 'f_favail',
@@ -108,58 +123,88 @@ class Fusecry(Operations):
     @debug_log
     def unlink(self, path):
         real_path = self.__real_path(path)
+        if real_path == self.conf: return None
         return os.unlink(real_path)
     
     @debug_log
     def symlink(self, name, target):
-        self.__real_path(path)
-        return os.symlink(target, real_path)
+        real_name = self.__real_path(name)
+        real_target = self.__real_path(target)
+        if real_name == self.conf: return None
+        if real_target == self.conf: return None
+        return os.symlink(real_target, real_name)
     
     @debug_log
     def rename(self, old, new):
-        return os.rename(self.__real_path(old), self.__real_path(new))
+        real_old = self.__real_path(old)
+        real_new = self.__real_path(new)
+        if real_old == self.conf: return None
+        if real_new == self.conf: return None
+        return os.rename(real_old, real_new)
     
     @debug_log
     def link(self, target, name):
-        return os.link(self.__real_path(name), self.__real_path(target))
+        real_target = self.__real_path(target)
+        real_name = self.__real_path(name)
+        if real_target == self.conf: return None
+        if real_name == self.conf: return None
+        return os.link(real_name, real_target)
     
     @debug_log
     def utimens(self, path, times=None):
-        return os.utime(self.__real_path(path), times)
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None
+        return os.utime(real_path, times)
 
     # File methods
     # ============
     
     @debug_log
     def open(self, path, flags):
-        return os.open(self.__real_path(path), flags)
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None
+        return os.open(real_path, flags)
     
     @debug_log
     def create(self, path, mode, fi=None):
-        return os.open(self.__real_path(path), os.O_WRONLY | os.O_CREAT, mode)
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None
+        return os.open(real_path, os.O_WRONLY | os.O_CREAT, mode)
 
     @debug_log
     def read(self, path, length, offset, fh):
-        return self.io.read(self.__real_path(path), length, offset)
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None, False
+        return self.io.read(real_path, length, offset)
 
     @debug_log
     def write(self, path, buf, offset, fh):
-        return self.io.write(self.__real_path(path), buf, offset)
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None
+        return self.io.write(real_path, buf, offset)
     
     @debug_log
     def truncate(self, path, length, fh=None):
-        return self.io.truncate(self.__real_path(path), length)
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None
+        return self.io.truncate(real_path, length)
     
     @debug_log
     def flush(self, path, fh):
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None
         return os.fsync(fh)
     
     @debug_log
     def release(self, path, fh):
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None
         return os.close(fh)
     
     @debug_log
     def fsync(self, path, fdatasync, fh):
+        real_path = self.__real_path(path)
+        if real_path == self.conf: return None
         return self.flush(path, fh)
